Amad

+++ app/src/main/java/com/tugas/aplikasimonitoringgudang/api/ApiSupplierService.kt
@@ -10,17 +10,17 @@ interface ApiSupplierService {
     suspend fun getSupplier(): SupplierResponse
 
     @GET("/api/supplier/{id}")
-    suspend fun getSupplierById(@Path("id") id: Int): SupplierResponseDetail
+    suspend fun getSupplierById(@Path("id") id: Long): SupplierResponseDetail
 
     @POST("/api/supplier")
     suspend fun addSupplier(@Body supplier: Supplier): SupplierResponseDetail
 
     @PUT("/api/supplier/{id}")
     suspend fun updateSupplier(
-        @Path("id") id: Int,
+        @Path("id") id: Long,
         @Body supplier: Supplier
     ): SupplierResponseDetail
 
     @DELETE("/api/supplier/{id}")
-    suspend fun deleteSupplier(@Path("id") id: Int): SupplierResponseDetail
+    suspend fun deleteSupplier(@Path("id") id: Long): SupplierResponseDetail
 }
